chore(discover): Cleanup events tests
- Delete the deprecated eventsv2 tests
- Move MEP tests to its own file
